feature: add db funcs and type assert

+++ app/domain/model/transaction.go
@@ -3,8 +3,8 @@ package model
 import "time"
 
 type Transaction struct {
-	ID              int
-	AccountID       int
+	ID              int64
+	AccountID       int64
 	OperationTypeID OperationType
 	Amount          float64
 	EventDate       time.Time
