+++ app/server/rest/model/request/findaccount.go
@@ -1,5 +1,5 @@
 package request
 
 type FindAccount struct {
-	AccountID int `json:"-" param:"accountId" validate:"required,gt=0"`
+	AccountID int64 `json:"-" param:"accountId" validate:"required,gt=0"`
 }
