+++ app/server/rest/model/response/transaction.go
@@ -7,9 +7,9 @@ import (
 )
 
 type Transaction struct {
-	ID              int       `json:"transaction_id"`
-	AccountID       int       `json:"account_id"`
-	OperationTypeID int       `json:"operation_type_id"`
+	ID              int64     `json:"transaction_id"`
+	AccountID       int64     `json:"account_id"`
+	OperationTypeID int64     `json:"operation_type_id"`
 	Amount          float64   `json:"amount"`
 	EventDate       time.Time `json:"event_date"`
 }
